chore: refactor generation of invite-codes_local.json
